Fix bug on response-output-coercion when using :stream
When :stream was being used on clj-http.client/request an
error of casting was being thrown. I added a check when
coercion is :stream to don't do anything to it (the same
way :byte-array does). I also added a test that would fail
without the code that was added on clj-http.client.