Fix wrap-nested-params for non-form content-types
The `wrap-nested-params` always kicked in, except when the :content-type
was set to :json. When sending nested structures as :edn or :transit the
payload got also nested. This PR fixes that problem, by only nesting if
there is no content type set (which defaults to x-www-form-urlencoded)
or if the content type is x-www-form-urlencoded.