Separate Transit reader and writer options
As mentioned in issue #292, the Transit reader and writer require
different options. At the moment `:transit-opts` is passed to both
reader and writer, which means you can either specify read handlers or
write handlers, but not both. Transit will complain if `:handlers` is a
mix of read and write handlers.

This patch splits up `:transit-opts` into a `:encode` and `:decode`
section. The options in `:encode` will be passed to the writer, the
options in `:decode` to the reader.