Use the canonical example domains as examples
The documentation and tests had been using a variety of domain names
other than the canonical example domains. These domains are
inappropriate and incorrect for this usage because they can be (and, in
practice, are) owned and operated by unknown parties who could respond
as in unpredictable ways to requests. These domains are especially
inappropriate and incorrect for sample code in documentation, because
any sample code presented will be copy-pasted and used, resulting in
users knocking on the door of unknown parties' servers.

This patch amends all instances of the following domains to instead use
IETF-reserved canonical example domains (or, in one case, to use the
".invalid" TLD, which is reserved for similar use: see RFC 6761 and RFC
2606).

* site dot com (owned by Salesforce)
* site dot come (typo for site dot com)
* foo dot com (owned by a domain name squatter)
* bar dot com (a Wordpress blog about the person who used to operate the
  email address "foo at bar dot com")
* media dot com (owned by "Future Media Architects, Inc")
* aoeu dot com (a placeholder site that encourages visitors to use the
  canonical example domains )
* some-site dot com (a Japanese SEO spammer)
* tmblr dot co (Tumblr's URL shortening domain)
* aoeuntahuf89o dot com (currently available for purchase)
* aorecuf892983a dot com (currently available for purchase)