Add NTLM Authentication example to the README
CHANGES
========
- Add an example of NTLM authentication to the readme.

WHY
========
When searching for a solution to NTLM + Clojure all I found was the issues and
PRs for clj-http. It would have been great to see it somewhere in the docs so I
knew if it worked or not. Hopefully this will begin driving traffic to the
readme of this project rather than the issues and PRs.