Wrap nested querystring params before form params
When encoding form params, a mutative change to the `:content-type` parameter is
made. This can cause nested query string parameters to be encoded incorrectly.

This commit changes the query string parameters (nested ones at least) to be
encoding prior to the form parameter coercion, and adds a test for the expected
behavior.