Fix redirects to an empty location
In some cases a server can respond with a redirect to an invalid location (like
"https:///"). With an async request, this leads to the request hanging forever.

This adds an option to do rudimentary validation of the redirect host in the
redirect stragety itself, with an option (`:validate-redirects`) to turn it off.

Resolves #416