Handle empty gzipped responses
When a request returns a response that is empty in addition to the
"content-encoding: gzip" header, clj-http would try to decode the response,
causing an EOFException.

This resolves this by attempting to read a single byte, and returning an empty
byte array if there is no body. This also fixes the case where the body was
attempted to be coerced and failed due to the `EOFException`

Resolves #257