Add support for re-using HttpClient and HttpAsyncClient
This commit adds the `:http-client` parameter in the response. This is the
HttpClient that was used for the request. An http client can also be passed in
with `:http-client`.

Note that an :http-client may only be specified when a reusable connection
manager is used, otherwise an exception will be thrown.

This also adds docs and naming clarification for the helpers in the `core`
namespace to create regular and async http clients.

Resolves #441