Redo 'with-async-connection-pool'
This commit drastically changes the way `with-async-connection-pool` works.
Previously we relied on creating a fake client and and starting to start the
io-reactor, and used a custom `:pooling-info` or `*pooling-info*` to do
poor man's reference counting for shutting down the pool.

No longer, this is now simplified and relies on the user managing scope
themselves.

Resolves #443