Strip quotes in detect-charset
RFC 7231 Section 3.1.1.1 allows the charset specified in the
Content-Type header to be a quoted string. We want detect-charset to
return unquoted charset names, so that they can be passed to, say,
java.nio.charset.Charset/forName without postprocessing.

This commit fixes that, adds a test for a quoted charset, and
rewrites t-detect-charset-by-content-type to use clojure.test/are
to reduce repetition and read more nicely.