fix(client): Handle multiple layers of compression
If a server responds with a `Content-Encoding` of something like
`deflate, gzip`, clj-http would not be able to handle the response as
the response body has been compressed using zlib and then compressed
again using gzip.

This patch adds a recursive wrapper to the decompression handling that
allows for multiple layers of compression to be unwrapped when receiving
responses like this.

The content encoding header is split by commas, then in reverse order is
decompressed layer by layer until the final body is produced which is
then returned to the caller.